[#406] NEW: clean and move impl of Automatic Driver Management
... to selene.core.configuration
- added shared.browser to selene.__init__.py

The impl. is still a draft, with broken entity.with(**config.as_kwargs) logic. TODO: make config to store "boxed objects"... so after applying .with(...) – the not overwritten options will be same, not new;)

+++ selene/core/wait.py
@@ -75,7 +75,7 @@ class Wait(Generic[E]):
     def __init__(
         self,
         entity: E,
-        at_most: int,
+        at_most: float,
         or_fail_with: Optional[Callable[[TimeoutException], Exception]] = None,
         _decorator: Callable[
             [Wait[E]], Callable[[fp.T], fp.T]
@@ -95,7 +95,7 @@ class Wait(Generic[E]):
         )
         return self.entity
 
-    def at_most(self, timeout: int) -> Wait[E]:
+    def at_most(self, timeout: float) -> Wait[E]:
         return Wait(self.entity, timeout, self._hook_failure)
 
     def or_fail_with(
